`click-fullscreen-control` renamed from `fullscreen-control`
