+++ src/test-utilities/click/click-fullscreen-control.ts
